refactor : 접근 제한자 수정, notnull 처리
