+++ src/main/java/com/gdsc/homework/domain/article/ArticleRepository.java
@@ -1,7 +1,6 @@
-package com.gdsc.homework.repository;
+package com.gdsc.homework.domain.article;
 
-import com.gdsc.homework.domain.Article;
-import com.gdsc.homework.domain.User;
+import com.gdsc.homework.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
