+++ src/main/java/com/gdsc/homework/service/ArticleService.java
@@ -1,8 +1,8 @@
 package com.gdsc.homework.service;
 
-import com.gdsc.homework.domain.Article;
-import com.gdsc.homework.domain.User;
-import com.gdsc.homework.repository.ArticleRepository;
+import com.gdsc.homework.domain.article.Article;
+import com.gdsc.homework.domain.user.User;
+import com.gdsc.homework.domain.article.ArticleRepository;
 import com.gdsc.homework.service.dto.request.ArticleDTO;
 import com.gdsc.homework.service.dto.response.ArticleResponse;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +10,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
